Split main.py into css_loader.py
